Add files via upload
Here is the interface(Buyable), abstract class(Item), subclass(Burger), object class(Spec), and driver. In this code, the connection between Burger and item has essentially been omitted because of an unknown error. To test this error, uncomment the methods in Item and comment out the methods in Burger. Once addressing the propper comments and uncomments, you should notice that Burger does not have access to the methods in Item for some unknown reason. The driver is used for testing.